uint64_t everywhere

+++ benchmarks/fcb-windows.c
@@ -1,5 +1,7 @@
+#include <_types/_uint64_t.h>
 #include <assert.h>
 #include <process.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <Windows.h>
 
@@ -10,15 +12,15 @@ struct benchmark_args {
   int sum_mod;
 };
 
-unsigned int sum_of_squares(int n)
+uint64_t sum_of_squares(uint64_t n)
 {
   return n*(n+1)*(2*n+1)/6;
 }
 
-unsigned int expected_value(int n_calls, int arg_mod, int sum_mod)
+uint64_t expected_value(uint64_t n_calls, uint64_t arg_mod, uint64_t sum_mod)
 {
-  int full_sum = ((n_calls / arg_mod) * sum_of_squares(arg_mod - 1)) % sum_mod;
-  int partial_sum = sum_of_squares((n_calls % arg_mod) - 1) % sum_mod;
+  uint64_t full_sum = ((n_calls / arg_mod) * sum_of_squares(arg_mod - 1)) % sum_mod;
+  uint64_t partial_sum = sum_of_squares((n_calls % arg_mod) - 1) % sum_mod;
 
   return (full_sum + partial_sum) % sum_mod;
 }
@@ -27,8 +29,8 @@ __stdcall
 unsigned int run_call_benchmark(void *argp)
 {
   struct benchmark_args *args = argp;
-  int sum = 0;
-  int (*fn)(int) = args->fn_ptr;
+  uint64_t sum = 0;
+  uint64_t (*fn)(uint64_t) = args->fn_ptr;
 
   LARGE_INTEGER start_ticks, end_ticks, ticks_per_sec;
   double elapsed_time;
@@ -36,7 +38,7 @@ unsigned int run_call_benchmark(void *argp)
   QueryPerformanceFrequency(&ticks_per_sec);
   QueryPerformanceCounter(&start_ticks);
 
-  for (int i = 0; i < args->n_calls; i++) {
+  for (uint64_t i = 0; i < args->n_calls; i++) {
     sum = (sum + fn(i % args->arg_mod)) % args->sum_mod;
   }
 
@@ -50,9 +52,9 @@ unsigned int run_call_benchmark(void *argp)
 }
 
 __declspec(dllexport)
-void benchmark_control(int n_calls, int arg_mod, int sum_mod)
+void benchmark_control(uint64_t n_calls, uint64_t arg_mod, uint64_t sum_mod)
 {
-  int sum = 0;
+  uint64_t sum = 0;
 
   LARGE_INTEGER start_ticks, end_ticks, ticks_per_sec;
   double elapsed_time;
@@ -60,8 +62,8 @@ void benchmark_control(int n_calls, int arg_mod, int sum_mod)
   QueryPerformanceFrequency(&ticks_per_sec);
   QueryPerformanceCounter(&start_ticks);
 
-  for (int i = 0; i < n_calls; i++) {
-    int x = i % arg_mod;
+  for (uint64_t i = 0; i < n_calls; i++) {
+    uint64_t x = i % arg_mod;
     sum = (sum + x*x) % sum_mod;
   }
 
@@ -74,13 +76,13 @@ void benchmark_control(int n_calls, int arg_mod, int sum_mod)
 }
 
 __declspec(dllexport)
-int square(int x)
+uint64_t square(uint64_t x)
 {
   return x * x;
 }
 
 __declspec(dllexport)
-void benchmark_calls_from_same_thread(void *fn_ptr, int n_calls, int arg_mod, int sum_mod)
+void benchmark_calls_from_same_thread(void *fn_ptr, uint64_t n_calls, uint64_t arg_mod, uint64_t sum_mod)
 {
   struct benchmark_args args;
 
@@ -93,7 +95,7 @@ void benchmark_calls_from_same_thread(void *fn_ptr, int n_calls, int arg_mod, in
 }
 
 __declspec(dllexport)
-__stdcall void benchmark_calls_from_new_thread(void *fn_ptr, int n_calls, int arg_mod, int sum_mod)
+__stdcall void benchmark_calls_from_new_thread(void *fn_ptr, uint64_t n_calls, uint64_t arg_mod, uint64_t sum_mod)
 {
   struct benchmark_args args;
